requirements.txt
streamlit
matplotlib
pandas
scikit-learn
numpy
mlflow
minio
python-dotenv
argparse
